Made various time-based values take double magnitude and retain full accuracy on converting to another time unit.
(cherry picked from commit 4bf92385f3430fb08093a5408aa76f218ebb0fbe)

+++ OcavaScenarioTest/src/main/java/com/ocadotechnology/scenario/ScenarioSimulationApi.java
@@ -26,7 +26,7 @@ import com.ocadotechnology.event.scheduling.EventScheduler;
 public interface ScenarioSimulationApi {
     boolean isStarted();
     EventScheduler getEventScheduler();
-    void setDiscreteEventTimeout(long duration, TimeUnit unit);
+    void setDiscreteEventTimeout(double duration, TimeUnit unit);
     void start(ScenarioNotificationListener listener);
     double getSchedulerStartTime();
     TimeUnit getSchedulerTimeUnit();
